Only return files
I only wish to return files, not directories, and highlight the filename
or the pattern that has been matched in the file.

This implements this by only printing lines if the FileInfo returns if
it is not a directory.

To highlight the filename, I split up the elements of the path, do a
replacement of the match for the name of the file only, and rejoin the
slice.